Update to PGA.m for compatibility with both MTEX 5.10 and 5.11
The release of MTEX 5.11.2 broke the PGA function due to syntax changes by the developers that did not maintain compatibility with previous versions. This new function includes a check for the MTEX version being used, and selects the appropriate syntax for proper computation.

Likely, these oversights by the MTEX developers will be adjusted for future releases, but currently this update to the code is necessary for proper computation in 5.11.2.